pyproject.toml
[tool.poetry]
name = "momuhackathon"
version = "0.1.0"
description = ""
authors = ["Ruben Peeters <40251899+RubenPeeters@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
rdflib = "*"
langchain = "*"
langchain-community = "^0.3.7"
python-dotenv = "^1.0.1"
langchain-experimental = "^0.3.3"
json-repair = "^0.30.2"
neo4j = "^5.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
